Transaction log - azure restores now match documentation
Added transaction logs and primary node doesn't commit until quorum is reached. There is an issue where a restoring nodecan receive a transaction and not apply it, the other replica is down, so the promary doesn't apply but that restoring replica will when it finishes restoring. Not sure what to do there